requirements.txt
pytest
pytest-html
py4j
pylint
selenium
python-dotenv
requests